// src/payload-types.ts
export interface Category {
  id: string;
  name: string;
  slug: string;
  color?: string | null;
  parent?: (string | null) | Category;
  subcategories?: {
    docs?: (string | Category)[];
    hasNextPage?: boolean;
    totalDocs?: number;
  };
  updatedAt: string;
  createdAt: string;
}
/**
 * You must verify your account before creating products.
 *
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "products".
 */
export interface Product {
  id: string;
  tenant?: (string | null) | Tenant;
  name: string;
  description?: string | null;
  /**
   * Price in INR
   */
  price: number;
  category?: (string | null) | Category;
  tags?: (string | Tag)[] | null;
  image?: (string | null) | Media;
  /**
   * Cover image for product detail page (larger view)
   */
  cover?: (string | null) | Media;
  refundPolicy?: ('30-day' | '14-day' | '7-day' | '3-day' | '1-day' | 'no-refunds') | null;
  /**
   * Protected content only visible to customers after purchase, Add product documentation, downloadable files, getting started guides and bonus materials. Supports rich-text formatting.
   */
  content?: string | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "tenants".
 */
export interface Tenant {
  id: string;
  /**
   * This is the name of the store you are creating. (e.g. John Doe's Store)
   */
  name: string;
  /**
   * This is the subdomain for the store you are creating. (e.g. [slug].xspark-ebazaar.com)
   */
  slug: string;
  /**
   * This is the image for the store you are creating.
   */
  image?: (string | null) | Media;
  /**
   * Stripe account ID associated with your store.
   */
  stripeAccountId: string;
  /**
   * You cannot create products until you have submitted your Stripe details.
   */
  stripeDetailsSubmitted?: boolean | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "media".
 */
export interface Media {
  id: string;
  alt: string;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "tags".
 */
export interface Tag {
  id: string;
  name: string;
  products?: (string | Product)[] | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "users".
 */
export interface User {
  id: string;
  username: string;
  roles?: ('user' | 'super-admin')[] | null;
  tenants?:
    | {
        tenant: string | Tenant;
        id?: string | null;
      }[]
    | null;
  updatedAt: string;
  createdAt: string;
  email: string;
  resetPasswordToken?: string | null;
  resetPasswordExpiration?: string | null;
  salt?: string | null;
  hash?: string | null;
  loginAttempts?: number | null;
  lockUntil?: string | null;
  sessions?:
    | {
        id: string;
        createdAt?: string | null;
        expiresAt: string;
      }[]
    | null;
  password?: string | null;
}
// ...
export interface Review {
  id: string;
  description: string;
  rating: number;
  product: string | Product;
  user: string | User;
  updatedAt: string;
  createdAt: string;
}